requirements.txt
pyasn1
python-nmap
python-snap7
scrapy
pysnmp
pylogix
pycomm3

